feat: add pydantic model for request and response
