+++ open_gpt/serve/gateway.py
@@ -1,154 +1,12 @@
 """The serve module provides a simple way to serve a model using Jina."""
 import json
-import logging
 
 import jina
 from jina import Document, DocumentArray
 from jina import Gateway as BaseGateway
 from jina.serve.runtimes.servers.composite import CompositeServer
-from pydantic import BaseModel, Field
-from typing import Union, List
-
-
-class GenerateRequest(BaseModel):
-    prompt: str = Field(..., description='The prompt to generate from.')
-
-    # session id
-    id: str = Field(description='The session id of the generation.', default=None)
-
-    # generation parameters
-    num_beams: int = Field(description='The number of beams to use.', default=None)
-    max_tokens: int = Field(
-        description='The maximum length of the generated text.', default=None
-    )
-    temperature: float = Field(
-        description='The temperature of the generation.', default=None
-    )
-    top_k: int = Field(description='The top k of the generation.', default=None)
-    top_p: float = Field(description='The top p of the generation.', default=None)
-    repetition_penalty: float = Field(
-        description='The repetition penalty of the generation.', default=None
-    )
-    logprobs: int = Field(
-        description='Include the log probabilities on the logprobs '
-        'most likely tokens, as well the chosen tokens',
-        default=None,
-    )
-    echo: bool = Field(
-        description='Echo back the prompt in the completion.', default=None
-    )
-    stop: Union[str, List[str]] = Field(description='Stop sequence generation on token.', default=None)
-    stop_str: Union[str, List[str]] = Field(description='Stop sequence generation on token.', default=None)
-    do_sample: bool = Field(
-        description='Whether to sample from the generation.', default=None
-    )
-    presence_penalty: float = Field(description='Positive values penalize new tokens based on whether they appear in '
-                                                'the text so far, increasing the likelihood to talk about new topics.',
-                                    default=0)
-    frequency_penalty: float = Field(description='Positive values penalize new tokens based on their existing '
-                                                 'frequency in the text so far, decreasing the likelihood to repeat '
-                                                 'the same line verbatim.',
-                                     default=0)
-    best_of: int = Field(description='Generates best_of completions server-side and returns the "best" (the one with '
-                                     'the highest log probability per token). Results cannot be streamed.',
-                         default=None)
-    n: int = Field(description='The number of sequences to return.', default=None)
-
-    class Config:
-        allow_population_by_field_name = True
-        arbitrary_types_allowed = True
-
-        schema_extra = {
-            'example': {
-                'prompt': 'Hello, my name is',
-                'id': '18d92585-7b66-4b7c-b818-71287c122c57',
-                'num_beams': 5,
-                'max_tokens': 50,
-                'temperature': 0.7,
-                'top_k': 50,
-                'top_p': 0.95,
-                'repetition_penalty': 1.0,
-                'echo': False,
-                'stop': '\n',
-                'do_sample': True,
-                'presence_penalty': 0.0,
-                'frequency_penalty': 0.0,
-                'best_of': 5,
-                'logprobs': None,
-                'n': 3,
-            }
-        }
-
-
-class ChatRequest(BaseModel):
-    messages: List[dict] = Field(..., description='The prompt to generate from.')
-
-    # session id
-    id: str = Field(description='The session id of the generation.', default=None)
-
-    # generation parameters
-    num_beams: int = Field(description='The number of beams to use.', default=None)
-    max_tokens: int = Field(
-        description='The maximum length of the generated text.', default=None
-    )
-    temperature: float = Field(
-        description='The temperature of the generation.', default=None
-    )
-    top_k: int = Field(description='The top k of the generation.', default=None)
-    top_p: float = Field(description='The top p of the generation.', default=None)
-    repetition_penalty: float = Field(
-        description='The repetition penalty of the generation.', default=None
-    )
-    logprobs: int = Field(
-        description='Include the log probabilities on the logprobs '
-                    'most likely tokens, as well the chosen tokens',
-        default=None,
-    )
-    echo: bool = Field(
-        description='Echo back the prompt in the completion.', default=None
-    )
-    stop: Union[str, List[str]] = Field(description='Stop sequence generation on token.', default=None)
-    stop_str: Union[str, List[str]] = Field(description='Stop sequence generation on token.', default=None)
-    do_sample: bool = Field(
-        description='Whether to sample from the generation.', default=None
-    )
-    presence_penalty: float = Field(description='Positive values penalize new tokens based on whether they appear in '
-                                                'the text so far, increasing the likelihood to talk about new topics.',
-                                    default=0)
-    frequency_penalty: float = Field(description='Positive values penalize new tokens based on their existing '
-                                                 'frequency in the text so far, decreasing the likelihood to repeat '
-                                                 'the same line verbatim.',
-                                     default=0)
-    best_of: int = Field(description='Generates best_of completions server-side and returns the "best" (the one with '
-                                     'the highest log probability per token). Results cannot be streamed.',
-                         default=None)
-    n: int = Field(description='The number of sequences to return.', default=None)
-
-    class Config:
-        allow_population_by_field_name = True
-        arbitrary_types_allowed = True
-
-        schema_extra = {
-            'example': {
-                'messages': [{"role": "system", "content": "You are a helpful assistant."},
-                             {"role": "user", "content": "Hello!"}],
-                'id': '18d92585-7b66-4b7c-b818-71287c122c57',
-                'num_beams': 5,
-                'max_tokens': 50,
-                'temperature': 0.7,
-                'top_k': 50,
-                'top_p': 0.95,
-                'repetition_penalty': 1.0,
-                'echo': False,
-                'stop': '\n',
-                'do_sample': True,
-                'presence_penalty': 0.0,
-                'frequency_penalty': 0.0,
-                'best_of': 5,
-                'logprobs': None,
-                'n': 3,
-            }
-        }
+from .pydantic_models import GenerateRequest, ChatRequest, BaseResponse
+from fastapi.encoders import jsonable_encoder
 
 
 class Gateway(BaseGateway, CompositeServer):
@@ -162,17 +20,6 @@ class Gateway(BaseGateway, CompositeServer):
         from fastapi import Body, status
         from fastapi.responses import JSONResponse
 
-        def _update_key(parameters):
-            key_maps = {
-                'max_tokens': 'max_new_tokens',
-                'n': 'num_return_sequences',
-                'stop': 'stop_str',
-            }
-            for openai_key, hf_key in key_maps.items():
-                if openai_key in parameters:
-                    parameters[hf_key] = parameters.pop(openai_key)
-            return parameters
-
         def _extend_rest_function(app):
             @app.api_route(path='/generate', methods=['POST'])
             @app.api_route(path='/codegen/completions', methods=['POST'])
@@ -185,8 +32,6 @@ class Gateway(BaseGateway, CompositeServer):
                     exclude_defaults=True,
                     exclude={'prompt'},
                 )
-
-                parameters = _update_key(parameters)
 
                 async for docs, error in self.streamer.stream(
                     docs=DocumentArray(
@@ -210,7 +55,7 @@ class Gateway(BaseGateway, CompositeServer):
 
                         return JSONResponse(
                             status_code=status.HTTP_200_OK,
-                            content=_tags,
+                            content=jsonable_encoder(BaseResponse(**_tags)),
                         )
 
             @app.api_route(path='/generate_stream', methods=['POST'])
@@ -226,8 +71,6 @@ class Gateway(BaseGateway, CompositeServer):
                     exclude_defaults=True,
                     exclude={'prompt'},
                 )
-
-                parameters = _update_key(parameters)
 
                 async def event_generator():
                     completion_tokens = 0
@@ -264,7 +107,7 @@ class Gateway(BaseGateway, CompositeServer):
                             _tags['usage'] = {
                                 k: int(v) for k, v in _tags['usage'].items()
                             }
-                            yield {"data": json.dumps(_tags)}
+                            yield {"data": jsonable_encoder(BaseResponse(**_tags))}
 
                 input_docs = DocumentArray(
                     [
@@ -285,8 +128,6 @@ class Gateway(BaseGateway, CompositeServer):
                     exclude_defaults=True,
                     exclude={'messages'},
                 )
-
-                parameters = _update_key(parameters)
 
                 async for docs, error in self.streamer.stream(
                     docs=DocumentArray(
@@ -310,7 +151,7 @@ class Gateway(BaseGateway, CompositeServer):
 
                         return JSONResponse(
                             status_code=status.HTTP_200_OK,
-                            content=_tags,
+                            content=jsonable_encoder(BaseResponse(**_tags)),
                         )
 
             @app.api_route(path='/chat_stream', methods=['POST'])
@@ -326,8 +167,6 @@ class Gateway(BaseGateway, CompositeServer):
                     exclude_defaults=True,
                     exclude={'messages'},
                 )
-
-                parameters = _update_key(parameters)
 
                 async def event_generator():
                     completion_tokens = 0
@@ -364,7 +203,7 @@ class Gateway(BaseGateway, CompositeServer):
                             _tags['usage'] = {
                                 k: int(v) for k, v in _tags['usage'].items()
                             }
-                            yield {"data": json.dumps(_tags)}
+                            yield {"data": jsonable_encoder(BaseResponse(**_tags))}
 
                 input_docs = DocumentArray(
                     [
